Create Get-Vulnerable-Systems-w-OutofDate-AVSignature
This query helps the operator find systems in MS Defender ATP that are vulnerable to exploitation due to not having one of the latest Defender Antivirus Signature Versions. Microsoft updates Defender Signature Versions multiple times a day, so having one of the latest version is crucial to keeping systems protected. Special credit goes to MJMELONE , who assisted me with the development of this query.